% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dig.R
\name{dig}
\alias{dig}
\title{Dig into a JSON data frame.}
\usage{
dig(jsondata, keyname, rs = FALSE)
}
\arguments{
\item{jsondata}{\code{data.frame} Data frame output from reading a JSON array data into R using \link[jsonlite]{\code{jsonlite::fromJSON}}.}

\item{keyname}{\code{character} A key name to dig into \code{jsondata}.}

\item{rs}{\code{logical} Short for return summary. Should \code{dig} return the data or a summary of the data?}
}
\description{
Dig into a JSON data frame.
}
\examples{
library(jsonlite)
library(dplyr)
jsondata <- fromJSON("https://api.github.com/users/hadley/repos", flatten = FALSE)
jsondata \%>\% dig("owner") \%>\% dig("type")
jsondata \%>\% dig("owner", rs= T)
jsondata \%>\% dig("owner") \%>\% summary.doc()
}
