% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{fetch}
\alias{fetch}
\title{Print a high-level summary of a data frame from JSON data}
\usage{
fetch(jsondata)
}
\arguments{
\item{jsondata}{\code{data.frame} Data frame output from reading a JSON array data into R using \code{\link[jsonlite]{jsonlite::fromJSON}}}
}
\description{
Print a high-level summary of a data frame from JSON data
}
\examples{
jsondata <- fromJSON("https://api.github.com/users/hadley/repos")
jsondata \%>\% fetch

}
