% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_doc.R
\name{summary.doc}
\alias{summary.doc}
\title{Print a high-level summary of a data frame from JSON data}
\usage{
\method{summary}{doc}(jsondata)
}
\arguments{
\item{jsondata}{\code{data.frame} Data frame output from reading a JSON array data into R using \link[jsonlite]{\code{jsonlite::fromJSON}}}
}
\description{
Print a high-level summary of a data frame from JSON data
}
\examples{
jsondata <- fromJSON("https://api.github.com/users/hadley/repos")
jsondata \%>\% summary.doc

}
